Update PulseShapeFitOOTPileupCorrection.cc
This update only effects the code for Method 2 in the case that the original double-pulse fitting is used. An if statement has been removed somewhere along the line of updates to this code, causing it to incorrectly assign rechit time & energy for certain cases. I have restored a suitable statement.  This option will probably not be used, but I think this should be fixed anyway.